Object value fields: Introduce m_int32Values.  * encapsulate basic value fields.  * resolved new warnings.

+++ src/scripts/SpellHandlers/WarlockSpells.cpp
@@ -581,7 +581,7 @@ bool DemonicKnowledge(uint8_t effectIndex, Aura* a, bool apply)
             else
                 val = (int32)-dmginc;
 
-            for (uint32 x = 0; x < 7; x++)
+            for (uint16_t x = 0; x < 7; x++)
                 PetOwner->ModPosDamageDoneMod(x, val);
 
             PetOwner->CalcDamage();
